Convert core code to library
